--- src/components/Input.jsx
@@ -1,7 +1,0 @@
-export default function Input({
-  children, onChange, onKeyPress, placeholder,
-}) {
-  return (
-    <input value={children} onChange={onChange} onKeyPress={onKeyPress} placeholder={placeholder} />
-  );
-}
