+++ src/components/common/Input.jsx
@@ -1,0 +1,12 @@
+export default function Input({
+  children, onChange, onKeyPress, placeholder,
+}) {
+  return (
+    <input
+      value={children}
+      onChange={onChange}
+      onKeyPress={onKeyPress}
+      placeholder={placeholder}
+    />
+  );
+}
